Add Error handling for notify and request calls
This changes the signature of notify() and request() to return a Result,
which will indicate if request can be made or not. For notify, the
operation may fail if the inbound channel is full. For request, the
operation may fail if the inbound channel is full or if there are no
available signals to capture the result.

Unit tests that verify Actor behavior is added.

Default features now include "std", which is usually the common case for
tests and simple examples.